[scanner] refactor: split oversized source files (auto-qa)
Split three source files that exceeded the 500-line limit into focused
sibling modules. All public exports are preserved via barrel re-exports,
so external callers keep working unchanged.

Splits:

- src/components/cards/cardMetadata.ts (775 -> 10 barrel)
  * cardTitles.ts        CARD_TITLES
  * cardDescriptions.ts  CARD_DESCRIPTIONS
  * cardDemoExempt.ts    DEMO_EXEMPT_CARDS

- src/lib/llmd/benchmarkMockData.ts (684 -> 50 barrel)
  * benchmarkMockData.types.ts       interfaces
  * benchmarkMockData.generators.ts  constants + generators
  * benchmarkMockData.pareto.ts      Pareto / leaderboard analysis

- src/hooks/mcp/types.ts (701 -> 366 core + barrel)
  * types.workloads.ts   Pod/Container/Deployment/Job/HPA/RS/SS/DS/CronJob
  * types.gpu.ts         AcceleratorType/GPUNode/GPUOperator/etc.
  * types.resources.ts   ConfigMap/Secret/SA/PVC/PV/Quota/LimitRange/RBAC

Also updates src/test/card-loading-standard.test.ts to exempt the new
sibling data files (cardTitles.ts, cardDemoExempt.ts) from the
card-loading convention check, matching the existing exemption pattern
for cardDescriptions.ts / cardMetadata.ts.

Issue: kubestellar/console#20967 (auto-QA oversized source files)
