+++ web/src/components/cards/cardDescriptions.ts
@@ -1,0 +1,356 @@
+// Card description registry — pure data, no runtime deps.
+
+// Short descriptions shown via info icon tooltip in the card header
+export const CARD_DESCRIPTIONS: Record<string, string> = {
+  acmm_level: "The repo's current level on the AI Codebase Maturity Model (L1–L6).",
+  acmm_feedback_loops: 'Inventory of criteria from ACMM, Fullsend, AEF, and Claude-Reflect.',
+  acmm_recommendations: 'Your current role and prioritized missing criteria for the next level.',
+  cluster_health: 'Overall health status of all connected Kubernetes clusters.',
+  cluster_focus: 'Deep-dive view of a single cluster with key metrics and resources.',
+  cluster_network: 'Network connectivity and traffic flow between clusters.',
+  cluster_comparison: 'Side-by-side comparison of clusters by resource usage and health.',
+  cluster_costs: 'Estimated infrastructure costs broken down by cluster.',
+  cluster_metrics: 'Real-time CPU, memory, and pod metrics across clusters.',
+  cluster_locations: 'Geographic map of cluster locations worldwide.',
+  cluster_resource_tree: 'Hierarchical tree view of all resources in a cluster.',
+  app_status: 'Status of workloads across clusters with health indicators.',
+  workload_deployment: 'Deploy workloads to clusters using drag-and-drop.',
+  deployment_missions: 'Track multi-cluster deployment missions and their progress.',
+  deployment_progress: 'Real-time deployment rollout progress and status.',
+  deployment_status: 'Detailed status of deployments including replicas and conditions.',
+  deployment_issues: 'Active deployment problems such as failed rollouts or image pull errors.',
+  statefulset_status: 'Status of StatefulSets including replicas and volume claims.',
+  daemonset_status: 'Status of DaemonSets including node coverage and update progress.',
+  replicaset_status: 'Status of ReplicaSets including replica counts and conditions.',
+  job_status: 'Status of Jobs including completion, duration, and failures.',
+  cronjob_status: 'Status of CronJobs including schedules and recent runs.',
+  hpa_status: 'Horizontal Pod Autoscalers (HPA) automatically adjust the number of pod replicas based on CPU, memory, or custom metrics. This card shows which workloads have autoscaling configured and whether they are currently scaling up or down.',
+  cluster_groups: 'Organize clusters into logical groups for targeted deployments.',
+  resource_marshall: 'Explore resource dependency trees and ownership chains.',
+  workload_monitor: 'Monitor all resources for a workload with health status, alerts, and AI diagnose/repair.',
+  llmd_stack_monitor: 'Monitor the llm-d inference stack: model serving, EPP, gateways, and autoscalers.',
+  prow_ci_monitor: 'Monitor PROW CI jobs with success rates, failure analysis, and AI repair.',
+  github_ci_monitor: 'Monitor GitHub Actions workflows across repos with pass rates and alerts.',
+  nightly_release_pulse: 'Last release tag, success/failure streak, next scheduled nightly, and a 14-day history strip.',
+  workflow_matrix: 'Heatmap of every workflow run over the last 14, 30, or 90 days — surfaces chronic flakiness.',
+  pipeline_flow: 'Drasi-style flow visualization of in-flight GitHub Actions runs: trigger, workflow, jobs, steps.',
+  recent_failures: 'The last failed GitHub Actions runs with log drill-down and optional re-run.',
+  cluster_health_monitor: 'Monitor cluster health across all connected clusters with pod and deployment issues.',
+  pod_issues: 'Pods with errors, restarts, or scheduling problems.',
+  top_pods: 'Top resource-consuming pods ranked by CPU or memory usage.',
+  resource_capacity: 'Cluster resource capacity vs. current allocation.',
+  resource_usage: 'CPU and memory allocation breakdown across clusters.',
+  node_status: 'Status of Kubernetes nodes including conditions and capacity.',
+  compute_overview: 'Summary of compute resources: nodes, CPUs, and memory.',
+  event_stream: 'Live stream of Kubernetes events from all clusters.',
+  event_summary: 'Aggregated event counts grouped by type and reason.',
+  warning_events: 'Warning-level events that may need attention.',
+  recent_events: 'Most recent events across all clusters.',
+  events_timeline: 'Timeline chart of event frequency over time.',
+  pod_logs: 'Live tail of container logs for any pod across your clusters.',
+  pod_health_trend: 'Historical trend of pod health status over time.',
+  resource_trend: 'Resource usage trends showing CPU and memory over time.',
+  storage_overview: 'PVC and storage class overview across clusters.',
+  pvc_status: 'Status of Persistent Volume Claims across clusters.',
+  pv_status: 'Status of Persistent Volumes including capacity and binding.',
+  resource_quota_status: 'Resource quota utilization and limits across namespaces.',
+  network_overview: 'Network policies, services, and ingress summary.',
+  service_status: 'Status of Kubernetes services and their endpoints.',
+  ingress_status: 'Status of Ingress resources including hosts and backends.',
+  network_policy_status: 'Status of Network Policies and affected pods.',
+  namespace_overview: 'Summary of resources within a namespace.',
+  namespace_analysis: 'Detailed analysis of namespace health and resource usage.',
+  namespace_rbac: 'RBAC roles and bindings within a namespace.',
+  namespace_quotas: 'Resource quota utilization within a namespace.',
+  namespace_events: 'Events filtered to a specific namespace.',
+  namespace_monitor: 'Real-time monitoring of namespace resource trends.',
+  operator_status: 'Status of installed Kubernetes operators.',
+  operator_subscriptions: 'Operator subscriptions and update channels.',
+  operator_subscription_status: 'Detailed status of Operator Lifecycle Manager subscriptions.',
+  crd_health: 'Custom Resource Definitions (CRDs) extend Kubernetes with new resource types. This card shows which CRDs are installed, whether they are serving correctly, and if any have version conflicts or schema issues.',
+  configmap_status: 'Status of ConfigMaps including size and update times.',
+  gitops_drift: 'Detects when your live cluster state has drifted from what is defined in Git. GitOps tools like Flux or ArgoCD should keep clusters in sync with Git — this card highlights any discrepancies that may indicate manual changes or sync failures.',
+  helm_release_status: 'Status of Helm releases across clusters.',
+  helm_releases: 'List of all deployed Helm releases.',
+  helm_history: 'Revision history and rollback options for Helm releases.',
+  helm_values_diff: 'Diff of Helm values between revisions.',
+  kustomization_status: 'Status of Kustomize overlays and their resources.',
+  flux_status: 'Flux Git repositories, kustomizations, and Helm release reconciliation status across clusters.',
+  chaos_mesh_status: 'Chaos Mesh experiment status and workflow progress.',
+  overlay_comparison: 'Compare Kustomize overlays across environments.',
+  chart_versions: 'Available Helm chart versions and update status.',
+  argocd_applications: 'ArgoCD application inventory and sync status.',
+  argocd_applicationsets: 'ArgoCD ApplicationSets manage fleet-wide deployments using generators to produce Applications across clusters.',
+  argocd_sync_status: 'Sync status of ArgoCD-managed applications.',
+  argocd_health: 'Health of ArgoCD applications and components.',
+  gpu_overview: 'Summary of GPU resources across all clusters. Use this to see how many GPUs are available, allocated, and idle across your fleet — essential for AI/ML workload planning.',
+  gpu_status: 'Current GPU utilization and health status. Shows real-time metrics for each GPU device so you can spot overloaded or underused GPUs.',
+  gpu_inventory: 'Inventory of GPU nodes with model, memory, and driver info. Helps you understand what GPU hardware is available (e.g., A100, H100) and whether drivers are up to date.',
+  gpu_workloads: 'Workloads running on GPU-enabled nodes. See which pods and jobs are consuming GPU resources and how much each is using.',
+  gpu_utilization: 'Real-time GPU utilization percentage and temperature. Monitor whether GPUs are being fully utilized or sitting idle, and catch thermal issues early.',
+  gpu_usage_trend: 'Historical GPU usage trends over time. Identify usage patterns and plan capacity by seeing how GPU demand changes throughout the day or week.',
+  gpu_namespace_allocations: 'GPU allocation breakdown by namespace across clusters. Understand which teams or projects are consuming GPU resources to optimize sharing and cost.',
+  gpu_inventory_history: 'Historical view of GPU allocation over time. Uses metrics snapshots captured every 10 minutes to show trends in GPU usage, helping you plan capacity and spot usage patterns.',
+  gpu_node_health: 'Proactive GPU node health monitoring with device disappearance detection. Alerts you when GPUs vanish from nodes — a common hardware issue that requires power cycling to fix.',
+  hardware_health: 'Detects hardware device disappearances (GPUs, NICs, NVMe, InfiniBand) that often require a power cycle to recover. Common with SuperMicro/HGX systems. Also shows full device inventory per node.',
+  security_issues: 'Security vulnerabilities and misconfigurations detected.',
+  rbac_overview: 'Overview of RBAC roles, bindings, and permissions.',
+  policy_violations: 'Active policy violations from OPA, Kyverno, or other engines.',
+  opa_policies: 'OPA (Open Policy Agent) Gatekeeper enforces policies on Kubernetes resources — for example, blocking containers running as root or requiring resource limits. This card shows which policies are active and whether any are being violated.',
+  kyverno_policies: 'Kyverno is a Kubernetes-native policy engine that can validate, mutate, and generate resources. This card shows active policies, their enforcement mode (audit vs enforce), and violation counts.',
+  intoto_supply_chain: 'in-toto is a CNCF supply chain security framework that verifies every step of your software delivery pipeline. This card shows active layouts, step verification status, and any failed or missing attestations across your clusters.',
+  falco_alerts: 'Falco is a runtime security tool that detects unexpected behavior in your containers (e.g., shell access, file modifications, network connections). This card shows recent alerts and their severity.',
+  trestle_scan: 'Compliance Trestle (CNCF Sandbox) performs OSCAL-based compliance assessments against standards like NIST 800-53 and FedRAMP. This card shows which security controls pass or fail across your clusters.',
+  trivy_scan: 'Trivy scans container images for known vulnerabilities (CVEs), misconfigurations, and exposed secrets. This card shows scan results with severity levels so you can prioritize fixes.',
+  kubescape_scan: 'Kubescape scans your Kubernetes clusters against security frameworks like NSA/CISA and MITRE ATT&CK. This card shows your security posture score and highlights areas that need hardening.',
+  iso27001_audit: 'Interactive ISO 27001 compliance checklist with 70 security controls for Kubernetes clusters. Walk through each control to assess your organization\'s information security management practices.',
+  compliance_score: 'Overall compliance score across security frameworks. Aggregates results from tools like OPA, Kyverno, Trivy, and Kubescape to give you a single score showing how well your clusters meet security and compliance standards.',
+  runtime_attestation: 'Composite 0-100 trust score per cluster derived from four CNCF signals: image provenance (TUF), workload identity (SPIFFE/SPIRE), policy compliance (Kyverno), and runtime privilege posture. Click a cluster to see the per-signal breakdown and non-compliant workloads.',
+  vault_secrets: 'HashiCorp Vault is an external secrets manager that securely stores API keys, passwords, and certificates. This card shows which Vault secrets are synced into Kubernetes and whether any are failing to sync.',
+  external_secrets: 'External Secrets Operator syncs secrets from external providers (AWS Secrets Manager, Azure Key Vault, GCP Secret Manager) into Kubernetes Secrets. This card shows sync status and any errors.',
+  cert_manager: 'cert-manager automates TLS certificate provisioning and renewal using providers like Let\'s Encrypt. This card shows certificate status, expiry dates, and whether renewals are succeeding.',
+  // Cross-cluster compliance cards
+  fleet_compliance_heatmap: 'Grid view of compliance status across clusters by tool category.',
+  compliance_drift: 'Detects clusters deviating from fleet compliance baseline.',
+  cross_cluster_policy_comparison: 'Side-by-side policy comparison across selected clusters.',
+  recommended_policies: 'AI-powered policy gap analysis with one-click deployment.',
+  // active_alerts — registered via unified descriptor system
+  alert_rules: 'Configured alert rules and their evaluation status.',
+  opencost_overview: 'Cost allocation data from OpenCost.',
+  kubecost_overview: 'Cost breakdown and optimization from Kubecost.',
+  service_exports: 'Service exports make a Kubernetes service available for discovery by other clusters in a multi-cluster federation. This card shows which services are exported and their readiness status.',
+  service_imports: 'Service imports bring remote services from other clusters into the local cluster, enabling cross-cluster communication. This card shows imported services and whether they are resolving correctly.',
+  gateway_status: 'Gateway API resource status and routing.',
+  service_topology: 'Visual topology of service-to-service communication.',
+  // Cluster admin cards
+  predictive_health: 'Predictive health analysis using historical trends to forecast issues.',
+  node_debug: 'Interactive node debugging with logs, events, and resource inspection.',
+  control_plane_health: 'Health status of Kubernetes control plane components.',
+  node_conditions: 'Detailed node conditions including disk pressure, memory, and network.',
+  admission_webhooks: 'Status of admission webhooks and their configurations.',
+  dns_health: 'DNS resolution health and CoreDNS pod status.',
+  etcd_status: 'Etcd cluster health, leader status, and database size.',
+  network_policies: 'Network policy coverage and affected pods.',
+  rbac_explorer: 'Interactive RBAC role and binding explorer.',
+  maintenance_windows: 'Scheduled maintenance windows and their status.',
+  cluster_changelog: 'Recent changes to cluster resources and configurations.',
+  change_timeline: 'Cross-cluster change events visualized as a scatter plot over time.',
+  quota_heatmap: 'Resource quota utilization heatmap across namespaces.',
+
+  // Kagenti AI Agent Platform
+  kagenti_status: 'Overall Kagenti AI agent platform status.',
+  kagenti_agent_fleet: 'Fleet view of all deployed Kagenti AI agents.',
+  kagenti_build_pipeline: 'Kagenti agent build pipeline status and history.',
+  kagenti_lifecycle_manager: 'Unified lifecycle view of Kagenti agents and builds with activity timeline.',
+  kagenti_tool_registry: 'Registry of tools available to Kagenti agents.',
+  kagenti_agent_discovery: 'Discover and browse available Kagenti agents.',
+  kagenti_security: 'Security status and access controls for Kagenti agents.',
+  kagenti_security_posture: 'Overall security posture of the Kagenti platform.',
+  kagenti_topology: 'Topology view of Kagenti agent connections and dependencies.',
+
+  // Kagent CRD Dashboard
+  kagent_status: 'Overview of kagent agents, tools, and models.',
+  kagent_agent_fleet: 'All kagent agents across clusters.',
+  kagent_agent_list: 'Agent status monitoring with health metrics.',
+  kagent_lifecycle_state: 'Visual lifecycle state machine showing agent transitions (Pending, Accepted, Ready, Failed).',
+  kagent_tool_registry: 'ToolServer and RemoteMCPServer resources.',
+  kagent_model_providers: 'ModelConfig and ModelProviderConfig resources.',
+  kagent_agent_discovery: 'Agent A2A config, tools, and skills.',
+  kagent_security: 'Agent approval settings and tool permissions.',
+  kagent_topology: 'Agent-Tool-Model dependency graph.',
+
+  // Crossplane
+  crossplane_managed_resources: 'Crossplane lets you provision and manage cloud infrastructure (databases, buckets, networks) using Kubernetes APIs. This card shows your managed resources and whether they are in sync with the desired state.',
+
+  // Cloud Native Buildpacks
+  buildpacks_status: 'Cloud Native Buildpacks automatically detect your app language and build OCI container images without writing a Dockerfile. This card shows build status, image history, and builder versions.',
+
+  upgrade_status: 'Kubernetes version upgrade status and available upgrades.',
+  user_management: 'Manage console users and their roles.',
+  github_activity: 'Recent GitHub activity: commits, PRs, and issues.',
+  issue_activity_chart: 'Daily chart of issues opened vs closed and PRs merged, with configurable lookback period.',
+  kubectl: 'Interactive kubectl terminal for running commands.',
+  quality_dashboard: 'Real-time metrics on state integrity, AI bug sweeps, and auto-remediations.',
+  // weather — registered via unified descriptor system
+  rss_feed: 'RSS feed reader for Kubernetes news and blogs.',
+  iframe_embed: 'Embed an external web page inside a card.',
+  network_utils: 'Network diagnostic utilities: ping, DNS, traceroute.',
+  mobile_browser: 'Embedded mobile-sized browser for testing.',
+  console_ai_issues: 'AI-detected issues and recommended fixes.',
+  console_ai_kubeconfig_audit: 'AI audit of kubeconfig files for security and cleanup.',
+  console_ai_health_check: 'AI-powered cluster health analysis.',
+  console_ai_offline_detection: 'Monitors cluster health and predicts failures before they happen. Detects offline nodes, GPU exhaustion, resource pressure, and groups issues by root cause for efficient remediation.',
+  // stock_market_ticker — registered via unified descriptor system
+  prow_jobs: 'PROW CI/CD job status and results.',
+  prow_status: 'Overall PROW system health and queue depth.',
+  prow_history: 'Historical PROW job runs and success rates.',
+  llm_inference: 'llm-d is a Kubernetes-native platform for serving Large Language Models. This card shows inference endpoint health, request throughput, and latency metrics for your LLM deployments.',
+  llm_models: 'LLM models deployed via llm-d with version info. See which models are loaded, their parameter sizes, and which pods are serving them.',
+  model_endpoint_health: 'Health summary for llm-d model serving endpoints, including ready, degraded, and replica counts across discovered serving deployments.',
+  epp_health: 'Health summary for llm-d Endpoint Picker Pod deployments, including ready, degraded, unavailable, and replica status.',
+  llmd_flow: 'Animated visualization showing how an inference request flows through the llm-d stack: from the load balancer, to the Endpoint Picker Pod (EPP), to prefill and decode pods. Helps you understand the request lifecycle.',
+  llmd_ai_insights: 'AI-generated insights about llm-d performance, bottlenecks, and optimization recommendations. Uses your cluster metrics to suggest improvements like adjusting batch sizes or scaling pods.',
+  llmd_configurator: 'Configure llm-d deployment parameters interactively: replicas, autoscaling thresholds, model variants, and GPU resource limits. Changes can be applied directly to your cluster.',
+  kvcache_monitor: 'KV cache stores attention key-value pairs to avoid recomputation during LLM inference. This card monitors cache utilization, hit rates, and memory usage across inference pods to help optimize throughput.',
+  epp_routing: 'The Endpoint Picker Pod (EPP) routes inference requests to the optimal pod based on KV cache affinity — sending requests to pods that already have relevant context cached. This card visualizes routing decisions.',
+  pd_disaggregation: 'Prefill/Decode disaggregation separates LLM inference into two phases: prefill (processing the full prompt) and decode (generating tokens one at a time). This architecture uses separate pod pools optimized for each phase.',
+  ml_jobs: 'Machine learning training and batch job status.',
+  ml_notebooks: 'Jupyter notebook server status and resource usage.',
+  provider_health: 'Health and status of AI and cloud infrastructure providers.',
+  strimzi_status: 'Strimzi runs Apache Kafka on Kubernetes for event streaming and messaging. This card shows Kafka cluster health, topic status, and consumer group lag — helping you spot message backlogs before they cause issues.',
+  // Flatcar Container Linux
+  flatcar_status: 'Flatcar Container Linux node status, version info, and update readiness.',
+  // Artifact Hub
+  artifact_hub_status: 'Artifact Hub package discovery and repository sync status.',
+  // Fluentd log collector
+  fluentd_status: 'Fluentd log collector pod health, buffer status, and throughput.',
+  // Lima VM
+  lima_status: 'Lima VM instance status, resource usage, and configuration.',
+  // OpenFeature feature-flag management
+  openfeature_status: 'OpenFeature feature flags: provider status (flagd, LaunchDarkly, ...), flag definitions by type (boolean/string/number/json), and evaluation metrics.',
+  // OpenKruise advanced workloads
+  openkruise_status: 'OpenKruise advanced workload status (CloneSet, Advanced StatefulSet/DaemonSet) and SidecarSet injection.',
+
+  // Benchmark cards
+  nightly_e2e_status: 'Nightly end-to-end test results and pass/fail trends.',
+  benchmark_hero: 'Latest benchmark results with key performance metrics.',
+  pareto_frontier: 'Performance vs cost Pareto frontier for hardware configurations.',
+  hardware_leaderboard: 'Hardware configuration rankings by throughput and latency.',
+  latency_breakdown: 'Latency breakdown by stage: prefill, decode, and network.',
+  throughput_comparison: 'Throughput comparison across hardware and model configurations.',
+  performance_timeline: 'Performance metrics over time with regression detection.',
+  resource_utilization: 'Resource utilization efficiency across benchmark runs.',
+
+  // Games
+  sudoku_game: 'Classic Sudoku puzzle game with multiple difficulty levels.',
+  match_game: 'Memory matching game with Kubernetes resource icons.',
+  solitaire: 'Classic Klondike solitaire card game.',
+  checkers: 'Play checkers against an AI opponent.',
+  game_2048: 'Slide and merge tiles to reach 2048.',
+  kubedle: 'Wordle-style game with Kubernetes terminology.',
+  pod_sweeper: 'Minesweeper clone with a Kubernetes pod theme.',
+  container_tetris: 'Classic Tetris with container-shaped blocks.',
+  flappy_pod: 'Navigate a pod through cluster obstacles.',
+  kube_man: 'Pac-Man style game collecting resources in a cluster maze.',
+  kube_kong: 'Donkey Kong inspired platformer with Kubernetes theme.',
+  pod_pitfall: 'Pitfall-style adventure game as a pod.',
+  node_invaders: 'Space Invaders clone defending your cluster.',
+  pod_crosser: 'Frogger-style game crossing cluster traffic.',
+  pod_brothers: 'Super Mario Bros inspired platformer.',
+  kube_kart: 'Racing game through Kubernetes infrastructure.',
+  kube_pong: 'Classic Pong game with cluster theming.',
+  kube_snake: 'Snake game collecting Kubernetes resources.',
+  kube_galaga: 'Galaga-style shooter defending against threats.',
+  kube_doom: 'First-person debugging adventure.',
+  kube_chess: 'Chess game with Kubernetes-themed pieces.',
+  missile_command: 'Missile Command arcade game defending your cluster.',
+  kube_bert: 'Q*bert style platformer navigating Kubernetes pyramids.',
+  // CoreDNS
+  coredns_status: 'CoreDNS pod health, restart counts, and cluster status across clusters.',
+  // Backstage developer portal (CNCF incubating)
+  backstage_status: 'Backstage developer portal — app replicas, catalog entity inventory (Components/APIs/Systems/Domains/Resources/Users/Groups), plugin status, scaffolder templates, and last catalog sync.',
+  // Contour ingress proxy
+  contour_status: 'Contour ingress proxy status, HTTPProxy resources, and Envoy fleet health.',
+  // Dapr distributed application runtime
+  dapr_status: 'Dapr control plane health, Dapr-enabled application count, and configured components (state store / pub-sub / binding).',
+  // Envoy proxy (service mesh / edge)
+  envoy_status: 'Envoy Proxy listener health, upstream cluster health, and request/connection stats.',
+  // gRPC services (network / service communication)
+  grpc_status: 'gRPC service serving status, per-service RPS, p99 latency, and error rates.',
+  // Linkerd service mesh
+  linkerd_status: 'Linkerd service mesh meshed pods, success rate, RPS, and p99 latency per deployment.',
+  // Longhorn distributed block storage (CNCF Incubating)
+  longhorn_status: 'Longhorn distributed block storage: volumes (state/robustness), node status, replica health, and capacity utilization.',
+  // OpenFGA fine-grained authorization (CNCF Sandbox)
+  openfga_status: 'OpenFGA fine-grained authorization: stores, authorization models, relationship tuples, API throughput (Check/Expand/ListObjects), and latency percentiles.',
+  // OpenTelemetry collector
+  otel_status: 'OpenTelemetry Collectors: pipeline health, receivers and exporters, dropped telemetry, and export errors across connected clusters.',
+  // Rook cloud-native storage orchestrator (Ceph)
+  rook_status: 'Rook-managed CephClusters: Ceph health, OSD/MON/MGR counts, capacity usage, and PG state summary.',
+  // SPIFFE workload identity (CNCF graduated)
+  spiffe_status: 'SPIFFE/SPIRE workload identity: trust domain, SVID counts (x509/JWT), federated trust domains, and registration entries.',
+  // CNI (Container Network Interface) plugin status
+  cni_status: 'Container Network Interface plugin status: active plugin, per-node CNI readiness, pod network CIDR, and NetworkPolicy coverage across services.',
+  // SPIRE (SPIFFE Runtime Environment) — workload identity
+  spire_status: 'SPIRE workload identity — server pod health, agent DaemonSet coverage, attested agents, registration entries, and trust bundle age.',
+  // TiKV distributed key-value store
+  tikv_status: 'TiKV distributed key-value store: store nodes, region counts, leader counts, and capacity utilization across the cluster.',
+  // TUF (The Update Framework) repository metadata
+  tuf_status: 'TUF repository role metadata — root, targets, snapshot, timestamp — versions, expirations, thresholds, and signing status.',
+  // Vitess distributed MySQL
+  vitess_status: 'Vitess distributed MySQL: keyspaces, shards, tablets (PRIMARY/REPLICA/RDONLY), and replication lag.',
+  // wasmCloud WebAssembly lattice (CNCF incubating)
+  wasmcloud_status: 'wasmCloud lattice status: hosts, actors, capability providers, and link definitions across connected clusters.',
+  // Volcano batch/HPC scheduler (CNCF Incubating)
+  volcano_status: 'Volcano batch/HPC scheduler: queues, jobs (pending/running/completed/failed), pod groups, and GPU allocation.',
+  // CRI-O container runtime
+  crio_status: 'CRI-O container runtime metrics, image pulls, and pod sandbox status.',
+  // Cloud Custodian rules engine (CNCF incubating)
+  cloud_custodian_status: 'Cloud Custodian policy runs — success / fail / dry-run counts, last run per policy, top resources acted on, and violations by severity.',
+  // Containerd container runtime
+  containerd_status: 'Containerd runtime — running containers, image, namespace, state, and uptime.',
+  // Cortex horizontally scalable Prometheus (CNCF incubating — marketplace#35)
+  cortex_status: 'Cortex (CNCF incubating) — horizontally scalable Prometheus: microservice health, active series, ingestion rate, query rate, and tenant count.',
+  // Dragonfly P2P image/file distribution
+  dragonfly_status: 'Dragonfly P2P image/file distribution — manager, scheduler, seed-peers, and per-node dfdaemon agents with active tasks and cache hit rate.',
+
+  // KubeVela application delivery (CNCF Incubating)
+  kubevela_status: 'KubeVela Application CRs: workflow step progress, component + trait counts, and vela-core controller health.',
+  // CloudEvents monitoring
+  cloudevents_status: 'CloudEvents message flow, event source tracking, and delivery status.',
+
+  // Multi-cluster insights cards
+  cross_cluster_event_correlation: 'Unified timeline showing correlated warning events across multiple clusters.',
+  cluster_delta_detector: 'Detects differences between clusters sharing the same workloads.',
+  cascade_impact_map: 'Visualizes how issues cascade across clusters over time.',
+  config_drift_heatmap: 'Cluster-pair matrix showing degree of configuration drift.',
+  resource_imbalance_detector: 'Detects CPU/memory utilization skew across the fleet.',
+  right_size_advisor: 'Per-cluster sizing verdicts — under-provisioned, right-sized, or over-provisioned — with actionable recommendations and a headroom buffer slider.',
+  restart_correlation_matrix: 'Detects horizontal (app bug) vs vertical (infra issue) restart patterns.',
+  deployment_rollout_tracker: 'Tracks deployment rollout progress across clusters.',
+  // KEDA
+  keda_status: 'KEDA (Kubernetes Event-Driven Autoscaling) automatically scales workloads based on external event sources like message queues, databases, or custom metrics. This card shows which workloads are being autoscaled, their current triggers, and queue depths.',
+  // Keycloak Identity & Access Management
+  keycloak_status: 'Keycloak is a CNCF-incubating open-source Identity and Access Management solution. This card monitors the Keycloak Operator health, realm status, active user sessions, and registered clients across your clusters.',
+  // OpenYurt edge computing
+  openyurt_status: 'OpenYurt extends Kubernetes to edge computing scenarios. This card monitors edge node pools, node autonomy status, and Raven gateway connectivity between edge and cloud clusters.',
+  // KServe model serving (kubestellar/console-marketplace#38)
+  kserve_status: 'KServe is a CNCF incubating model-serving platform on Kubernetes. This card monitors InferenceService readiness, predictor replica status, canary traffic split, and serving throughput / latency across clusters.',
+  // Knative serverless
+  knative_status: 'Knative is a CNCF graduated platform for serverless workloads on Kubernetes. This card monitors Knative Serving services, revision status, traffic routing, and Eventing broker health across your clusters.',
+  // Karmada multi-cluster orchestration
+  karmada_status: 'Karmada is a multi-cluster orchestration tool that propagates resources (Deployments, Services, etc.) across multiple clusters using placement policies. This card shows propagation status, member cluster health, and policy compliance.',
+  cubefs_status: 'CubeFS distributed file system health, volume status, and node topology',
+  harbor_status: 'Harbor registry projects, repositories, and vulnerability scan results',
+  deployment_risk_score: 'Correlates Argo CD sync status, Kyverno violations, and pod restart rates into a single 0-100 risk score per namespace so one glance replaces five dashboards.',
+  kuberay_fleet: 'KubeRay fleet monitoring — RayCluster, RayService, and RayJob status across all clusters',
+  slo_compliance: 'Tracks SLO compliance with configurable targets for latency, error rate, and availability. Shows error budget burn rate and per-cluster compliance indicators.',
+  failover_timeline: 'Forensic timeline of cross-region failover events detected from Karmada ResourceBinding status transitions. Shows cluster outages, binding rescheduling, and recovery events.',
+  trino_gateway: 'Discovers Trino coordinator, worker, and Trino Gateway pods across clusters. Shows per-cluster query health, gateway routing status, and worker distribution.',
+  // Fluid dataset caching
+  fluid_status: 'Fluid is a CNCF incubating cloud-native dataset orchestrator that accelerates data access for AI/Big Data workloads. This card monitors dataset caching status, runtime engine health, and data load progress across your clusters.',
+
+  // Inspektor Gadget
+  network_trace: 'Live network connection tracing via Inspektor Gadget eBPF.',
+  dns_trace: 'Live DNS query tracing via Inspektor Gadget eBPF.',
+  process_trace: 'Live process execution tracing via Inspektor Gadget eBPF.',
+  security_audit: 'Security audit using Inspektor Gadget eBPF-based runtime analysis.',
+
+  // Multi-tenancy
+  ovn_status: 'OVN-Kubernetes network status, User Defined Networks, and tenant isolation.',
+  kubeflex_status: 'KubeFlex controller status, control planes per tenant, and CP health.',
+  k3s_status: 'K3s lightweight Kubernetes server pods, agent connections, and cluster health.',
+  kubevirt_status: 'KubeVirt VM status across clusters — running, stopped, paused, migrating, and error states with per-cluster breakdown, CPU/memory allocation, and data-plane isolation.',
+  vcluster_status: 'Virtual cluster status across host clusters: running, paused, and failed vClusters with Kubernetes version info.',
+  multi_tenancy_overview: 'Aggregated view of tenant isolation across OVN, KubeFlex, K3s, and KubeVirt.',
+  tenant_isolation_setup: 'AI-powered multi-tenancy setup wizard with component detection and one-click configuration.',
+  tenant_topology: 'Interactive SVG topology of the KubeCon multi-tenancy architecture: KubeVirt VMs, K3s control planes, Layer-2/3 UDN networks, and KubeFlex controller with live status indicators.',
+
+  // Quantum computing
+  quantum_status: 'Live status of the quantum circuit executor: backend connection, execution mode, shot count, and last run time.',
+  quantum_control_panel: 'Configure and trigger quantum circuit execution — select backend, QASM file, shot count, and toggle loop mode.',
+  quantum_qubit_grid: 'Pixel grid of qubit measurement outcomes; each cell represents a qubit state across recent shots.',
+  quantum_circuit_viewer: 'Renders the active QASM circuit diagram showing gates and circuit depth from the last execution.',
+  quantum_histogram: 'Measurement result histogram displaying bit-string outcome counts from the most recent circuit execution.',
+}
+
