+++ web/src/hooks/mcp/types.ts
@@ -1,3 +1,12 @@
+// MCP hook types — cluster, node, service, namespace, operator, helm, security, and metric types.
+//
+// Domain-specific groups are split into sibling files (re-exported below) to keep each under the
+// max-lines limit:
+//   - types.workloads.ts  (Pod/Container/Deployment/Job/HPA/ReplicaSet/StatefulSet/DaemonSet/CronJob)
+//   - types.gpu.ts        (AcceleratorType/GPUNode/GPUOperator/etc.)
+//   - types.resources.ts  (ConfigMap/Secret/ServiceAccount/PVC/PV/ResourceQuota/LimitRange/RBAC)
+// Public exports are preserved — external callers keep importing from './types'.
+
 export interface ClusterInfo {
   name: string
   context: string
@@ -96,50 +105,6 @@ export interface ClusterHealth {
   errorMessage?: string
 }
 
-export interface ContainerInfo {
-  name: string
-  image: string
-  ready: boolean
-  state: 'running' | 'waiting' | 'terminated'
-  reason?: string
-  message?: string
-  gpuRequested?: number  // Number of GPUs requested by this container
-}
-
-export interface PodInfo {
-  name: string
-  namespace: string
-  cluster?: string
-  status: string
-  ready: string
-  restarts: number
-  age: string
-  node?: string
-  labels?: Record<string, string>
-  annotations?: Record<string, string>
-  containers?: ContainerInfo[]
-  // Resource requests (sum of all containers)
-  cpuRequestMillis?: number    // CPU request in millicores
-  cpuLimitMillis?: number      // CPU limit in millicores
-  memoryRequestBytes?: number  // Memory request in bytes
-  memoryLimitBytes?: number    // Memory limit in bytes
-  gpuRequest?: number          // Total GPU request
-  // Actual resource usage (from metrics API, if available)
-  cpuUsageMillis?: number      // Actual CPU usage in millicores
-  memoryUsageBytes?: number    // Actual memory usage in bytes
-  metricsAvailable?: boolean   // Whether metrics API data is available
-}
-
-export interface PodIssue {
-  name: string
-  namespace: string
-  cluster?: string
-  status: string
-  reason?: string
-  issues: string[]
-  restarts: number
-}
-
 export interface ClusterEvent {
   type: string
   reason: string
@@ -150,34 +115,6 @@ export interface ClusterEvent {
   count: number
   firstSeen?: string
   lastSeen?: string
-}
-
-export interface DeploymentIssue {
-  name: string
-  namespace: string
-  cluster?: string
-  replicas: number
-  readyReplicas: number
-  reason?: string
-  message?: string
-}
-
-export interface Deployment {
-  name: string
-  namespace: string
-  cluster?: string
-  status: 'running' | 'deploying' | 'failed'
-  replicas: number
-  readyReplicas: number
-  updatedReplicas: number
-  availableReplicas: number
-  progress: number
-  image?: string
-  age?: string
-  labels?: Record<string, string>
-  annotations?: Record<string, string>
-  reason?: string
-  message?: string
 }
 
 /**
@@ -211,117 +148,6 @@ export interface AggregatedMetricCompleteness {
   contributingClusters: string[]
   missingClusters: string[]
   isComplete: boolean
-}
-
-// AcceleratorType represents the category of accelerator
-export type AcceleratorType = 'GPU' | 'TPU' | 'AIU' | 'XPU'
-
-/**
- * Scheduling-gating taint on a GPU node (issue #8172).
- * Only taint effects that actually gate scheduling (`NoSchedule`, `NoExecute`)
- * are surfaced by the backend — `PreferNoSchedule` is advisory and is omitted.
- */
-export interface GPUTaint {
-  key: string
-  value?: string
-  effect: string  // 'NoSchedule' | 'NoExecute'
-}
-
-export interface GPUNode {
-  name: string
-  cluster: string
-  gpuType: string           // Display name of accelerator (e.g., "NVIDIA A100", "Intel Gaudi2", "IBM AIU")
-  gpuCount: number          // Number of accelerators
-  gpuAllocated: number      // Number of allocated accelerators
-  acceleratorType?: AcceleratorType  // GPU, TPU, AIU, or XPU
-  /** Scheduling-gating taints on the underlying node (issue #8172). */
-  taints?: GPUTaint[]
-  // Enhanced GPU info from NVIDIA GPU Feature Discovery
-  gpuMemoryMB?: number
-  gpuFamily?: string
-  cudaDriverVersion?: string
-  cudaRuntimeVersion?: string
-  migCapable?: boolean
-  migStrategy?: string
-  manufacturer?: string     // Manufacturer (NVIDIA, AMD, Intel, Google, IBM)
-}
-
-// GPU Node Health Check types (proactive monitoring)
-export interface GPUNodeHealthCheck {
-  name: string
-  passed: boolean
-  message?: string
-}
-
-export interface GPUNodeHealthStatus {
-  nodeName: string
-  cluster: string
-  status: 'healthy' | 'degraded' | 'unhealthy'
-  gpuCount: number
-  gpuType: string
-  checks: GPUNodeHealthCheck[]
-  issues: string[]
-  stuckPods: number
-  checkedAt: string
-}
-
-export interface GPUHealthCheckResult {
-  nodeName: string
-  status: 'healthy' | 'degraded' | 'unhealthy'
-  gpuCount?: number
-  checks: GPUNodeHealthCheck[]
-  issues: string[]
-}
-
-export interface GPUHealthCronJobStatus {
-  installed: boolean
-  cluster: string
-  namespace?: string
-  schedule?: string
-  tier: number
-  version: number
-  updateAvailable: boolean
-  lastRun?: string
-  lastResult?: string
-  nextRun?: string
-  canInstall: boolean
-  activeJobs: number
-  failedJobs: number
-  successJobs: number
-  lastResults?: GPUHealthCheckResult[]
-}
-
-// NVIDIA Operator Status types
-export interface OperatorComponent {
-  name: string
-  status: string
-  reason?: string
-}
-
-export interface GPUOperatorInfo {
-  installed: boolean
-  version?: string
-  state?: string
-  ready: boolean
-  components?: OperatorComponent[]
-  driverVersion?: string
-  cudaVersion?: string
-  namespace?: string
-}
-
-export interface NetworkOperatorInfo {
-  installed: boolean
-  version?: string
-  state?: string
-  ready: boolean
-  components?: OperatorComponent[]
-  namespace?: string
-}
-
-export interface NVIDIAOperatorStatus {
-  cluster: string
-  gpuOperator?: GPUOperatorInfo
-  networkOperator?: NetworkOperatorInfo
 }
 
 export interface NodeCondition {
@@ -394,152 +220,6 @@ export interface Service {
   annotations?: Record<string, string>
 }
 
-export interface Job {
-  name: string
-  namespace: string
-  cluster?: string
-  status: string // Running, Complete, Failed
-  completions: string
-  duration?: string
-  age?: string
-  labels?: Record<string, string>
-  annotations?: Record<string, string>
-}
-
-export interface HPA {
-  name: string
-  namespace: string
-  cluster?: string
-  reference: string
-  minReplicas: number
-  maxReplicas: number
-  currentReplicas: number
-  targetCPU?: string
-  currentCPU?: string
-  age?: string
-  labels?: Record<string, string>
-  annotations?: Record<string, string>
-}
-
-export interface ConfigMap {
-  name: string
-  namespace: string
-  cluster?: string
-  dataCount: number
-  age?: string
-  labels?: Record<string, string>
-  annotations?: Record<string, string>
-}
-
-export interface Secret {
-  name: string
-  namespace: string
-  cluster?: string
-  type: string
-  dataCount: number
-  age?: string
-  labels?: Record<string, string>
-  annotations?: Record<string, string>
-}
-
-export interface ServiceAccount {
-  name: string
-  namespace: string
-  cluster?: string
-  secrets?: string[]
-  imagePullSecrets?: string[]
-  age?: string
-  labels?: Record<string, string>
-  annotations?: Record<string, string>
-}
-
-export interface PVC {
-  name: string
-  namespace: string
-  cluster?: string
-  status: string
-  storageClass?: string
-  capacity?: string
-  accessModes?: string[]
-  volumeName?: string
-  age?: string
-  labels?: Record<string, string>
-  annotations?: Record<string, string>
-}
-
-export interface PV {
-  name: string
-  cluster?: string
-  status: string
-  capacity?: string
-  storageClass?: string
-  reclaimPolicy?: string
-  accessModes?: string[]
-  claimRef?: string
-  volumeMode?: string
-  age?: string
-  labels?: Record<string, string>
-}
-
-export interface ResourceQuota {
-  name: string
-  namespace: string
-  cluster?: string
-  hard: Record<string, string>  // Resource limits
-  used: Record<string, string>  // Current usage
-  age?: string
-  labels?: Record<string, string>
-  annotations?: Record<string, string>  // Reservation metadata
-}
-
-export interface ReplicaSet {
-  name: string
-  namespace: string
-  cluster?: string
-  replicas: number
-  readyReplicas: number
-  ownerName?: string
-  ownerKind?: string
-  age?: string
-  labels?: Record<string, string>
-}
-
-export interface StatefulSet {
-  name: string
-  namespace: string
-  cluster?: string
-  replicas: number
-  readyReplicas: number
-  status: string
-  image?: string
-  age?: string
-  labels?: Record<string, string>
-}
-
-export interface DaemonSet {
-  name: string
-  namespace: string
-  cluster?: string
-  desiredScheduled: number
-  currentScheduled: number
-  ready: number
-  status: string
-  age?: string
-  labels?: Record<string, string>
-}
-
-export interface CronJob {
-  name: string
-  namespace: string
-  cluster?: string
-  schedule: string
-  suspend: boolean
-  active: number
-  lastSchedule?: string
-  age?: string
-  labels?: Record<string, string>
-}
-
 export interface Ingress {
   name: string
   namespace: string
@@ -561,23 +241,6 @@ export interface NetworkPolicy {
   labels?: Record<string, string>
 }
 
-export interface LimitRangeItem {
-  type: string  // Pod, Container, PersistentVolumeClaim
-  default?: Record<string, string>
-  defaultRequest?: Record<string, string>
-  max?: Record<string, string>
-  min?: Record<string, string>
-}
-
-export interface LimitRange {
-  name: string
-  namespace: string
-  cluster?: string
-  limits: LimitRangeItem[]
-  age?: string
-  labels?: Record<string, string>
-}
-
 export interface MCPStatus {
   opsClient: {
     available: boolean
@@ -587,16 +250,6 @@ export interface MCPStatus {
     available: boolean
     toolCount: number
   }
-}
-
-export interface ResourceQuotaSpec {
-  cluster: string
-  name: string
-  namespace: string
-  hard: Record<string, string>
-  labels?: Record<string, string>
-  annotations?: Record<string, string>  // Reservation metadata
-  ensure_namespace?: boolean  // Auto-create namespace if it doesn't exist (GPU reservation flow)
 }
 
 export interface SecurityIssue {
@@ -647,39 +300,6 @@ export interface OperatorSubscription {
   cluster?: string
 }
 
-export interface K8sRole {
-  name: string
-  namespace?: string
-  cluster: string
-  isCluster: boolean
-  ruleCount: number
-}
-
-// K8s role binding type
-export interface K8sRoleBinding {
-  name: string
-  namespace?: string
-  cluster: string
-  isCluster: boolean
-  roleName: string
-  roleKind: string
-  subjects: Array<{
-    kind: 'User' | 'Group' | 'ServiceAccount'
-    name: string
-    namespace?: string
-  }>
-}
-
-// K8s service account type (for RBAC)
-export interface K8sServiceAccountInfo {
-  name: string
-  namespace: string
-  cluster: string
-  secrets?: string[]
-  roles?: string[]
-  createdAt?: string
-}
-
 export interface HelmRelease {
   name: string
   namespace: string
@@ -699,3 +319,48 @@ export interface HelmHistoryEntry {
   app_version: string
   description: string
 }
+
+// -- Re-exports from split sibling files --
+
+export type {
+  ContainerInfo,
+  PodInfo,
+  PodIssue,
+  DeploymentIssue,
+  Deployment,
+  Job,
+  HPA,
+  ReplicaSet,
+  StatefulSet,
+  DaemonSet,
+  CronJob,
+} from './types.workloads'
+
+export type {
+  AcceleratorType,
+  GPUTaint,
+  GPUNode,
+  GPUNodeHealthCheck,
+  GPUNodeHealthStatus,
+  GPUHealthCheckResult,
+  GPUHealthCronJobStatus,
+  OperatorComponent,
+  GPUOperatorInfo,
+  NetworkOperatorInfo,
+  NVIDIAOperatorStatus,
+} from './types.gpu'
+
+export type {
+  ConfigMap,
+  Secret,
+  ServiceAccount,
+  PVC,
+  PV,
+  ResourceQuota,
+  LimitRangeItem,
+  LimitRange,
+  ResourceQuotaSpec,
+  K8sRole,
+  K8sRoleBinding,
+  K8sServiceAccountInfo,
+} from './types.resources'
